feat: lazy load query results on the landing page
